fixing registration backend
