+++ lux/extensions/rest/backends/mixins.py
@@ -1,4 +1,5 @@
 import time
+from datetime import datetime, timedelta
 
 from pulsar import ImproperlyConfigured
 from pulsar.utils.pep import to_string
@@ -15,6 +16,19 @@ except ImportError:     # pragma    nocover
 
 
 class TokenBackendMixin:
+    """Mixin for token and session based authentication back-ends
+    """
+    _config = [
+        Parameter('SESSION_EXPIRY', 7*24*60*60,
+                  'Expiry for a session/token in seconds.')
+    ]
+
+    def session_expiry(self, request):
+        '''Expiry for a session or a token
+        '''
+        session_expiry = request.config['SESSION_EXPIRY']
+        if session_expiry:
+            return datetime.now() + timedelta(seconds=session_expiry)
 
     def api_sections(self, app):
         '''At the authorization router to the api
